requirements.txt
websockets>=10.0
mysql-connector-python>=8.0
aiomysql>=0.1.1
aiohttp>=3.8.0
python-dotenv>=0.19.0
pandas>=1.3.0
numpy>=1.21.0
typing-extensions>=4.0.0
pytest>=6.2.5
black>=21.12b0
pylint>=2.12.0